fix conditions MenusController,RestaurantsController et UsersController

+++ src/controllers/MenusController.ts
@@ -9,30 +9,31 @@ export class MenusController extends BaseEntity {
 
     async addMenu(req: Request, res: Response) {
         const restoMenu: string = req.body.menu;
-        const priceMenu: number = req.body.price;
+        const priceMenu: number = parseInt(req.body.price);
+        const verifMenu = await menuService.verifMenuByMenu(restoMenu);
 
 
-        if (restoMenu !== restoMenu.toString()) {
+        if (typeof restoMenu !== 'string' && typeof priceMenu !== 'number') {
             res.status(400).json({
                 status: "FAIL",
-                message: "le nom du menu doit être une chaine de caractère"
+                message: "le nom incrit et le prix saisie ne corresponde pas au format défini"
+            });
+            return;
+        } if (verifMenu) {
+            res.status(400).json({
+                status: "Fail",
+                message: "Menu déjà existant"
             });
             return;
         }
         try {
             const menu = await menuService.addMenu(restoMenu, priceMenu);
-            if (menu) {
-                res.status(201).json({
-                    status: "Create",
-                    messsage: "Nouveau menu",
-                    data: menu
-                })
-            } else {
-                res.status(404).json({
-                    status: "Fail",
-                    message: "Création impossible"
-                })
-            }
+
+            res.status(201).json({
+                status: "Create",
+                messsage: "Nouveau menu",
+                data: menu
+            })
         } catch (err) {
             res.status(500).json({
                 status: "fail",
@@ -62,6 +63,14 @@ export class MenusController extends BaseEntity {
 
     async deleteMenu(req: Request, res: Response) {
         const id: number = parseInt(req.params.id);
+        const verifDelMenuId = await menuService.selectMenuById(id);
+        if (!verifDelMenuId || typeof id !== 'number') {
+            res.status(400).json({
+                status: "Fail",
+                message: "Menu inexistant"
+            });
+            return;
+        }
         try {
             const delMenu = menuService.deleteMenu(id);
             res.status(200).json({
@@ -80,20 +89,28 @@ export class MenusController extends BaseEntity {
     }
 
     async putMenu(req: Request, res: Response) {
-        const restoId = parseInt(req.params.id);
-        const name: string = req.body.menu;
+        const id = parseInt(req.params.id);
+        const restoMenu: string = req.body.menu;
         const price: number = req.body.price;
+        const verifPutMenuId = await menuService.selectMenuById(id);
 
-
-        if (!name) {
-            res.status(200).json({
-                status: "impossible",
-                message: "Aucune modification détecté"
+        if (typeof restoMenu !== 'string' || typeof price !== 'number' || typeof id !== 'number') {
+            res.status(400).json({
+                status: "FAIL",
+                message: "le nom incrit , le prix et l'id saisie ne correspondent pas au format défini"
             });
             return;
+        } if (!verifPutMenuId) {
+            res.status(400).json({
+                status: "Fail",
+                message: "Menu inexistant"
+            });
+            return;
+
+
         }
         try {
-            const putMenu = await menuService.putMenu(restoId, name, price);
+            const putMenu = await menuService.putMenu(id, restoMenu, price);
             res.status(200).json({
                 status: "Success",
                 message: "Modification effectuée",
